Workaround for broken pagination of os-simple-tenant-usage API
In some scenarios, pagination of the os-simple-tenant-usage API fails to work
correctly. When provided with an instance UUID marker, the server will return a
response that includes usage for that instance, but only instances after that
marker should be included. This causes the client to loop forever, believing
there is more usage data to be retrieved. This affects the overview page in
horizon, which retrieves this usage data.

This change checks for the case when the marker is included in returned data,
and in that case returns an empty dict. This will prevent the client from
making further requests.